Adds base charm implementation
Adds the charm. At this point, it sets up the nginx service inside the
workload container and expects a few configuration options which are
to be set.

Adds the Dockerfile for the certbot-nginx image and instructions on how
to build it. The image is based on the nginx docker image, and has
certbot installed, which can be used to generate CA-verified certificates
by solving their HTTP challenges.