Adds certbot implementation
Adds service-hostname config option. Once set, the charm will then
generate a CA-verified certificate, and then register it as a
Kubernetes Secret. If the secret already exists, a new certificate
will not be generated.

In order to create the Kubernetes Secrets, the charm will have to
be trusted. Otherwise, the 403 error is encountered when trying
to list / create Secrets.