Adds the Ingress relation
Adds the ingress relation in metadata.yaml and adds the
nginx_ingress_integrator/v0/ingress library.

The charm now requires an ingress relation in order to become
active. That relation will be used to set up the ACME Challenge
Route which will be needed in order for the CA to verify our
ownership of the hostname we're creating the certificate for.

When the service-hostname config option is set and there is no
secret created for it, it will set the service-hostname in the
Ingress relation data, and defers. The related nginx-ingress-integrator
charm will then create the Ingress route for us. Afterwards,
we're checking that the Ingress is properly set up through a test file
in order to make sure that the CA will be able to reach its desired
files in order to complete its Challenge.