Updates the status to Waiting on certification creation
Currently, the charm remains in an Active state even when waiting
for the Ingress Route to be set up or generating a certificate, which
can be a bit to know when the generation process ended.

We're now setting up a Waiting status when generating certificates,
adding information about the current step as well. The charm will
become Active again after the process ended.