Create permanent route for the HTTP challenge
We're currently creating the route for the HTTP challenge whenever
we need to create / renew the certificate by updating the ingress
relation data.

However, the relation data is not updated automatically, only after
the current event handled finishes or is defered. Which means that
the renew_certificate action cannot create the needed route.

This updates the current behaviour to always have the route created,
solving the renew_certificate's issue.