Ensures that the nginx service is running when setting up ingress route
In some scenarios, Pebble may be running, but there are no layers configured
on it due to various reasons, which means that the the 'nginx' service may not
be running, which means that the subsequent ingress route checks will fail.

With this change, we're ensuring that the Pebble layer is being created and
the service will be restarted if it's not running.